requirements.txt
PyMuPDF>=1.23.0
python-pptx>=0.6.23
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0